requirements.txt
simulaqron
flask
requests
mysql-connector-python
pycrypto
pycryptodome
progressbar
